Refactor Spotify Service Bus: Move types and delete safeSpotifyApi
- Move `SafeSpotifyApi` and related types to `types/spotify.ts`.
- Delete `services/safeSpotifyApi.ts`.
- Update `services/spotifyPolling.ts` imports.
- Previously: Fixed lint errors, build failures, and removed wrapper functions.

This addresses the final architectural feedback by consolidating types and removing "zombie files".
